Address feedback

+++ test/shared/src/main/scala/zio/test/TestAspect.scala
@@ -343,7 +343,7 @@ object TestAspect extends TimeoutVariants {
    * use with flaky tests.
    */
   def flaky(n: Int): TestAspectAtLeastR[ZTestEnv with Annotations with Live] =
-    retry(Schedule.recurs(n.toLong))
+    retry(Schedule.recurs(n))
 
   /**
    * An aspect that runs each test on its own separate fiber.
@@ -441,7 +441,7 @@ object TestAspect extends TimeoutVariants {
    * is stable ("non-flaky"). Stops at the first failure.
    */
   def nonFlaky(n: Int): TestAspectAtLeastR[ZTestEnv with Annotations with Live] =
-    repeat(Schedule.recurs(n.toLong))
+    repeat(Schedule.recurs(n))
 
   /**
    * Constructs an aspect that requires a test to not terminate within the
